+++ source/functions.lua
@@ -63,48 +63,90 @@ end
 ---@param combat any Combat container from Details!
 ---@param bossID number ID of the boss for this encounter
 ---@param bossName string Name of the boss for this encounter
-function PuGCalculateBuffPoints(raidComp, combat, bossID, bossName)
+function PuGCheckPotionBuffs(raidComp, combat, bossID, bossName)
     local miscContainer = combat:GetContainer(DETAILS_ATTRIBUTE_MISC)
-    print("Calculating buff points for players...")
     for _, actor in miscContainer:ListActors() do
-        local actorBufffs = {
-            isWellFed = false,
-            isFlask = false,
-            isPotion = false,
-        }
-        local actorPoints = 0
+        local isActorPotionBuff = false
         if actor.buff_uptime_spells then
             local buffContainer = actor.buff_uptime_spells
             if buffContainer and buffContainer._ActorTable then
                 for spellID, spellInfo in pairs(buffContainer._ActorTable) do
                     if type(spellInfo) == "table" then
                         local spellName = C_Spell.GetSpellInfo(spellID).name or "Unknown Spell"
-                        if string.find(spellName, "Well Fed") then
-                            actorBufffs.isWellFed = true
-                        elseif  string.find(spellName, "Flask of") then
-                            actorBufffs.isFlask = true
-                        elseif  string.find(spellName, "Potion") then
-                            actorBufffs.isPotion = true
+                        if string.find(spellName, "Potion") then
+                            isActorPotionBuff = true
                         end
                     end
                 end
             end
-            for _, hasBuff in pairs(actorBufffs) do
-                if hasBuff then
-                    actorPoints = actorPoints + 1
-                end
-            end
-            if actorBufffs.isWellFed and actorBufffs.isFlask and actorBufffs.isPotion then
-                actorPoints = actorPoints + 1
-            end
-            if actorPoints > 0 then
+            if isActorPotionBuff then
                 for _, player in ipairs(raidComp) do
                     if StripRealm(player.name) == actor:GetOnlyName() then
-                        PuGAddPointEntry(player, bossID, bossName, "Food/Flask/Potion", actorPoints)
-                        if PuGAddonDebug then print(PuGAddonDebugPrefix .. player.name .. " has earned " .. actorPoints .. " for having food/flask buffs and/or using potions.") end
+                        PuGAddPointEntry(player, bossID, bossName, "Potion Used", 1)
+                        if PuGAddonDebug then print(PuGAddonDebugPrefix .. player.name .. " has earned " .. 1 .. " for using potions.") end
                         break
                     end
                 end
+            end
+        end
+    end
+end
+
+--- Loop through all players in the raid and check if they have the specified buff
+--- Returns a table with player names as keys and their buff points as values
+---@return table unitsWithBuffs A table containing player names and their buff points
+function PuGCalculateBuffPoints()
+    local unitsWithBuffs = {}
+    for i = 1, GetNumGroupMembers() do
+        local name, _, _, _, _, _, _, _, _, _, _ = GetRaidRosterInfo(i)
+        local unit = "raid" .. i
+        local unitPoints = 0
+        local unitBuffs = {
+            isWellFed = false,
+            isFlask = false,
+            isAugment = false,
+        }
+        for j = 1, 40 do
+            local aura = C_UnitAuras.GetAuraDataByIndex(unit, j, "HELPFUL")
+            if not aura then
+                break
+            end
+            local buffName = aura.name
+            if string.find(buffName, "Well Fed") then
+                unitBuffs.isWellFed = true
+            elseif string.find(buffName, "Flask of") then
+                unitBuffs.isFlask = true
+            elseif string.find(buffName, "Crystallization") then
+                unitBuffs.isAugment = true
+            end
+        end
+        for _, hasBuff in pairs(unitBuffs) do
+            if hasBuff then
+                unitPoints = unitPoints + 1
+            end
+        end
+        if unitBuffs.isWellFed and unitBuffs.isFlask and unitBuffs.isAugment then
+            unitPoints = unitPoints + 1
+        end
+        if unitPoints > 0 then
+            unitsWithBuffs[name] = unitPoints -- unitWithBuffs = {name = name, points = unitPoints}
+        end
+    end
+    return unitsWithBuffs
+end
+
+--- Awards points to players who have food, flask, or augment buffs
+---@param raidComp table Raid roster to assign points to (normal comp or progress comp)
+---@param bossID number ID of the boss for this encounter
+---@param bossName string Name of the boss for this encounter
+---@param unitsWithBuffs table A table containing player names and their buff points
+function PuGGiveBuffPoints(raidComp, bossID, bossName, unitsWithBuffs)
+    for playerName, points in pairs(unitsWithBuffs) do
+        for _, player in ipairs(raidComp) do
+            if StripRealm(player.name) == StripRealm(playerName) then
+                PuGAddPointEntry(player, bossID, bossName, "Food/Flask/Augment Rune", points)
+                if PuGAddonDebug then print(PuGAddonDebugPrefix .. player.name .. " has earned " .. points .. " for having food/flask/augment buffs.") end
+                break
             end
         end
     end